+++ eras/dijkstra/impl/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## 0.3.0.0
 
+* Deprecate `BHeaderView` in favour of `DijkstraEraBlockHeader` typeclass.
+  - Add `PerasCert`, `PerasKey` and `validatePerasCert` to `Dijkstra.BlockBody` (moved from core).
+  - Add `DijkstraEraBlockHeader` typeclass and the `DijkstraBbodySignal` GADT.
+  - Remove `PrevEpochNonceNotPresent` from `DijkstraBbodyPredFailure`.
 * Add `SubLedgerEnv` and `SubUtxowEnv`
 * Remove `OutputTooSmallUTxO` constructor from `DijkstraUtxoPredFailure`
 * Remove `SubOutputTooSmallUTxO` constructor from `DijkstraSubUtxoPredFailure`